requirements.txt
asgiref==3.2.10
certifi==2020.12.5
Django==3.1.4
django-crispy-forms==1.10.0
django-phonenumber-field==5.0.0
django-template-maths==0.2.0
olefile==0.46
phonenumbers==8.12.14
gunicorn==20.0.4
pillow==7.2.0
pytz==2020.1
reportlab==3.5.51
sqlparse==0.3.1
whitenoise==5.2.0
wincertstore==0.2

